requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.8.4
deepeval==3.9.5
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
execnet==2.1.2
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.14
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
pillow==12.1.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyfiglet==1.0.4
Pygments==2.20.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywin32==311
PyYAML==6.0.3
ragas==0.4.3
regex==2026.3.32
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
scikit-network==0.33.5
scipy==1.17.1
sentry-sdk==2.57.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlglot==26.31.1
sqlparse==0.5.5
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
